move dependency tests to new file more tests and formatting
